MedLedger_mk1/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.11.0
attrs==25.3.0
bitarray==3.1.1
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
ckzg==2.1.0
click==8.1.8
cryptography==44.0.2
cssselect2==0.8.0
cytoolz==1.0.1
dnspython==2.7.0
eth-account==0.13.5
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.6.1
eth-rlp==2.2.0
eth-typing==5.2.0
eth-utils==5.2.0
eth_abi==5.2.0
fastapi==0.115.11
fonttools==4.57.0
frozenlist==1.5.0
h11==0.14.0
hexbytes==1.3.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
joblib==1.4.2
MarkupSafe==3.0.2
motor==3.7.0
multidict==6.2.0
notification==0.2.1
numpy==2.2.5
pandas==2.2.3
parsimonious==0.10.0
pillow==11.2.1
propcache==0.3.0
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.10.6
pydantic_core==2.27.2
pydyf==0.11.0
PyJWT==2.10.1
pymongo==4.11.2
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.2
pyunormalize==16.0.0
regex==2024.11.6
requests==2.32.3
rlp==4.1.0
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tinyhtml5==2.0.0
toolz==1.0.0
types-requests==2.32.0.20250306
typing_extensions==4.12.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.3.0
uvicorn==0.34.0
weasyprint==65.1
web3==7.9.0
webencodings==0.5.1
websockets==13.1
yarl==1.18.3
zopfli==0.2.3.post1
